Made a bunch of changes. Parametrized the source ranges with view to how we'd want to handle them in the future. Source ranges as they stand now are one of: IFCONDITION, STMT, DECL, COMPOUNDSTMT, INITIALIZER.
The visitor methods return a vector of OffsetRanges (as opposed to returning a
single OffsetRange as was the case previously) and the helper methods (used to
print the sourceranges etc.)  have been updated appropriately. This was done
because certain AST elements introduce more than just one 'interesting' source
range. For example, an if statement has a source range corresponding to the
entire if-then-else, but also a source range corresponding to the
if-condition. This allows uniform treatment of all such cases.

Also, re-indented the code. The mix of tabs and spaces was driving me
crazy. A bunch of other minor changes too, which I'm probably
forgetting. Should've split up this commit, but need to rush off to 706. More
changes and commits to follow, once they are in an acceptable state.