Removed obvious errors and fallacies. Fixed the Search strategy and evaluation sections.
After having fixed the problems in the constraint generator which were keeping
us from trying larger (and more complex) files i ran a comparison of our
algorithm with ddmin (run on the same syntactic granularity). Not only did our
algorithm converge in half the time and using an order of magnitude fewer tests,
it's output was a third the size of the minimizing test case got by running
ddmin. This was done using randomized search because nothing else was
feasible. Even with randomized search the fact that even though we were able to
reduce the number of tests by a factor of 10 and yet reduce the time only by a
factor of 2 points to the need for a better inference system than the one we
currently have. I might try a database solution before writing one myself. If i
do write my own though i think imma use it to learn Haskell.